Add RP_ZERO board and allow I2C controller selection.
Before this commit, I attempted to do I2C communication and discovered
that the existing configuration selected the incorrect I2C controller
and GPIO pins for the Raspberry Pi Zero (though it was presumably
correct for a different Raspberry Pi board). It was trying to use the
I2C controller numbered 0 by the BCM2835 peripherals document (BSC0),
presumably on GPIO pins 0 and 1. For the RPi Zero, it needs to use the
I2C controller numbered 1 (BSC1) and configured on GPIO pins 2 and 3.
(The RPi Zero does not expose GPIO pins 0 and 1,
but it does expose 2 and 3.)

This allows the I2C controller choice to be made by either using
macro defines (ex: pass the `-DBCM2835_I2C_USE_I2C1` to the compiler
from within a Makefile), or by include'ing the
`$(CHIBIOS)/boards/RP_ZERO/board.mk`
file in the project's Makefile instead of
`$(CHIBIOS)/boards/RASPBERRYPI_MODB/board.mk`.